# Conflicts:
#	Communication/Packets/Incoming/Rooms/Action/MuteUserEvent.cs
#	Communication/Packets/Incoming/Rooms/Avatar/ApplySignEvent.cs
#	Communication/Packets/Incoming/Rooms/Chat/ChatEvent.cs
#	Communication/Packets/Incoming/Rooms/Chat/ShoutEvent.cs
#	Communication/Packets/Incoming/Rooms/Chat/WhisperEvent.cs
#	Communication/Packets/Incoming/Rooms/Engine/GetRoomEntryDataEvent.cs
#	Communication/Packets/Incoming/Rooms/Settings/SaveRoomSettingsEvent.cs
#	Communication/Packets/Incoming/Users/SetChatPreferenceEvent.cs
#	Communication/Packets/Incoming/Users/SetMessengerInviteStatusEvent.cs
#	Communication/Packets/Incoming/Users/SetUserFocusPreferenceEvent.cs
#	Communication/Packets/Incoming/Users/UpdateFigureDataEvent.cs

+++ HabboHotel/Rooms/RoomUserManager.cs
@@ -148,7 +148,7 @@ public class RoomUserManager
                 var square = _room.GetGameMap().GetRandomWalkableSquare();
                 model.DoorX = square.X;
                 model.DoorY = square.Y;
-                model.DoorZ = _room.GetGameMap().GetHeightForSquareFromData(square);
+                model.DoorZ = (int)_room.GetGameMap().GetHeightForSquareFromData(square);
             }
             user.SetPos(model.DoorX, model.DoorY, model.DoorZ);
             user.SetRot(model.DoorOrientation, false);
@@ -950,7 +950,7 @@ public class RoomUserManager
                                 user.Team = Team.None;
                             }
                             //Item.ExtraData = usersOnTeam.ToString();
-                            //Item.UpdateState(false, true);                                
+                            //Item.UpdateState(false, true);
                         }
                         break;
                     }
@@ -990,7 +990,7 @@ public class RoomUserManager
                                 user.Team = Team.None;
                             }
                             //Item.ExtraData = usersOnTeam.ToString();
-                            //Item.UpdateState(false, true);                                
+                            //Item.UpdateState(false, true);
                         }
                         break;
                     }
